Fix WoD garrison base, levels/building specifics still need to be imp

+++ src/server/game/Server/Packets/GarrisonPackets.cpp
@@ -480,8 +480,8 @@ WorldPacket const* WorldPackets::Garrison::GarrisonLandingPage::Write()
 
 WorldPacket const* WorldPackets::Garrison::GarrisonAddMissionResult::Write()
 {
-    _worldPacket << GarrTypeID;
-    _worldPacket << Result;
+    _worldPacket << int32(GarrTypeID);
+    _worldPacket << int32(Result);
     _worldPacket << unk; // 2 == GARRISON_RANDOM_MISSION_ADDED = 904
     _worldPacket << MissionData;
 
@@ -896,7 +896,7 @@ WorldPacket const* WorldPackets::Garrison::GarrisonOpenMissionNpc::Write()
     _worldPacket << Result;
     _worldPacket << static_cast<uint32>(Missions.size());
     for (auto const& missionID : Missions)
-        _worldPacket << missionID;
+        _worldPacket << int32(missionID);
 
     _worldPacket.WriteBit(UnkBit1);
     _worldPacket.WriteBit(PreventXmlOpenMissionEvent);
